+++ src/main/java/com/example/vvs/domain/common/s3/S3Config.java
@@ -1,4 +1,4 @@
-package com.example.vvs.domain.common;
+package com.example.vvs.domain.common.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
