+++ src/main/java/com/example/vvs/domain/common/s3/S3Service.java
@@ -1,4 +1,4 @@
-package com.example.vvs.domain.common;
+package com.example.vvs.domain.common.s3;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
